+++ apps/angular-lib-dev/src/app/form/form.component.ts
@@ -20,6 +20,7 @@ import '@sebgroup/green-core/components/rich-text/index.js'
 import '@sebgroup/green-core/components/text/index.js'
 import '@sebgroup/green-core/components/theme/index.js'
 import '@sebgroup/green-core/components/segmented-control/index.js'
+import '@sebgroup/green-core/components/select/index.js'
 
 import { registerTransitionalStyles } from '@sebgroup/green-core/transitional-styles'
 
@@ -45,6 +46,7 @@ export class FormComponent {
     email: ['', Validators.required],
     description: ['', Validators.required],
     dropdown: ['', Validators.required],
+    select: ['', Validators.required],
     date: [undefined as Date | undefined, Validators.required],
   })