+++ libs/angular/src/v-angular/tabs/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Public API Surface of tabs
+ */
+
+export * from './tab.directive'
+export * from './tabs.component'
+export * from './tabs.models'
+export * from './tabs.module'