+++ libs/core/src/components/popover/popover.styles.ts
@@ -14,7 +14,7 @@ const style = css`
     :host([open]) dialog {
       opacity: 1;
       box-sizing: border-box;
-      transform: translate3d(0, 0, 0);
+      /*transform: translate3d(0, 0, 0);*/
       visibility: visible;
     }
 
@@ -35,14 +35,52 @@ const style = css`
       border-radius: 0.5rem;
     }
 
+    @media (max-width: 767px) {
+      dialog.use-modal-in-mobile {
+        border-radius: 1rem 1rem 0 0;
+        transition: all 200ms ease;
+        min-width: 100vw;
+        position: fixed;
+        bottom: 0;
+        left: 0;
+
+        max-height: 50svh;
+        padding-bottom: 0;
+
+        transform: translateY(0);
+        @starting-style {
+          transform: translateY(100%);
+        }
+
+        &::backdrop {
+          background-color: rgba(0, 0, 0, 0.3);
+          display: block;
+          position: fixed;
+          opacity: 1;
+          transition: opacity 200ms ease;
+          @starting-style {
+            opacity: 0;
+          }
+        }
+      }
+    }
+
+    @media (min-width: 768px) {
+      header {
+        display: none;
+      }
+    }
+
     header {
       display: none;
     }
 
-    dialog:not(.has-backdrop)::backdrop {
-      background-color: transparent;
-      display: block;
-      position: fixed;
+    @media (min-width: 768px) {
+      dialog:not(.has-backdrop)::backdrop {
+        background-color: transparent;
+        display: block;
+        position: fixed;
+      }
     }
   }
 `